==> Entrega/Procesamiento/1-Carga.Rmd <==
---
title: "PREDICCION ACCIDENTES MEDELLIN"
author: "Edwin Villarraga-Jorge Montoya-Juan David Arango"
date: "10/8/2019"
output:
    html_document:
      theme: yeti
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
```

# Carga de Datos

Durante la carga de datos se realizo la verificacion de las gravedad de accidentes e inicialmente se identificaron 3 categorias:
1. Solo Daños
2. Herido
3. Muerto

Agrupamiento de los datos:

1. Accidentes Graves: Heridos y Muertos
2. Accidentes Leves: Solo Daños

## Datos de 2014


```{r Datos_2014}
raw_data_2014 <- read.csv(file="../data/Accidentalidad_georreferenciada_2014.csv", encoding="UTF-8", header=TRUE, sep=",")
```

## Datos de 2015

```{r Datos_2015}
raw_data_2015 <- read.csv(file="../data/Accidentalidad_georreferenciada_2015.csv", encoding="UTF-8", header=TRUE, sep=",")
```

## Datos 2016

```{r Datos_2016}
raw_data_2016 <- read.csv(file="../data/Accidentalidad_georreferenciada_2016.csv", encoding="UTF-8", header=TRUE, sep=",")
```


## Datos 2017

```{r Datos_2017}
raw_data_2017 <- read.csv(file="../data/Accidentalidad_georreferenciada_2017.csv", encoding="UTF-8", header=TRUE, sep=",")
```

## Datos 2018

```{r Datos_2018}
raw_data_2018 <- read.csv(file="../data/Accidentalidad_georreferenciada_2018.csv", encoding="UTF-8", header=TRUE, sep=",")
```


## Union de los datos

```{r}
Total_Dataset <- rbind(raw_data_2014,raw_data_2015,raw_data_2016, raw_data_2017, raw_data_2018)
```

```{r}
str(Total_Dataset)
```
```{r}
summary(Total_Dataset$GRAVEDAD)
```


```{r}
levels(Total_Dataset$GRAVEDAD)
```

Se agrupan las variables por el nivel de gravedad del accidente. Donde accidente grave es aquel incidente donde se presente afectación de individuos

```{r}
levels(Total_Dataset$GRAVEDAD) <- c("ACCIDENTES_GRAVES","ACCIDENTES_GRAVES","ACCIDENTES_LEVES")
```



```{r}
levels(Total_Dataset$GRAVEDAD)
```



Para ahorra memoria se eliminan los dataframes por ano

```{r}
rm(raw_data_2014,raw_data_2015,raw_data_2016, raw_data_2017, raw_data_2018)
```


La Fecha esta cargada como factor, se realiza una transformacion a formato de fecha (date)


```{r}
Total_Dataset$FECHA <- as.Date(Total_Dataset$FECHA, format="%Y-%m-%d")
```

```{r}
Total_Dataset <- Total_Dataset[,c('GRAVEDAD','FECHA','DIA','DIA_NOMBRE','MES')]
tail(Total_Dataset)
```

Se convierten las variables dia y mes a factor

```{r}
Total_Dataset$DIA <- as.factor(Total_Dataset$DIA)
Total_Dataset$MES <- as.factor(Total_Dataset$MES)
```


```{r}
tail(Total_Dataset)
```

Se genera frecuencia que es la tabla a analizar

```{r}
library(sqldf)
Total_Dataset_Freq <- sqldf("SELECT FECHA, GRAVEDAD, count(GRAVEDAD) AS FREQ, DIA_NOMBRE, MES, DIA 
       FROM Total_Dataset
       GROUP BY FECHA, GRAVEDAD")
```

```{r}
Total_Dataset_Freq$Fecha<- NULL
```


se agrega la variable ano al datafrate

```{r}
Total_Dataset_Freq$ANO <- as.factor(format(Total_Dataset_Freq$FECHA,'%Y'))
```

Se agrega la variable semana al dataframe

```{r}
Total_Dataset_Freq$SEMANA <-as.factor(format(Total_Dataset_Freq$FECHA,'%V'))
```

```{r}
Total_Dataset_Freq$MES <-as.factor(format(Total_Dataset_Freq$FECHA,'%m'))
```


```{r}
head(Total_Dataset_Freq)
```

Se depura la variable DIA_NOMBRE para eliminar los espacios, quitar tildes, dar un orden al factor

```{r}
levels(Total_Dataset_Freq$DIA_NOMBRE)
```

```{r}
levels(Total_Dataset_Freq$DIA_NOMBRE) <- c("DOMINGO","JUEVES","LUNES","MARTES","MIERCOLES","SABADO","VIERNES") 
```

```{r}
levels(Total_Dataset_Freq$DIA_NOMBRE)
```

Se corre esta linea para ordenar los niveles del factor día de la semana

```{r}
Total_Dataset_Freq$DIA_NOMBRE <- ordered(Total_Dataset_Freq$DIA_NOMBRE,c("LUNES", "MARTES", "MIERCOLES", "JUEVES", "VIERNES", "SABADO", "DOMINGO"))
```

```{r}
levels(Total_Dataset_Freq$DIA_NOMBRE)
```

Ahora se agregan otras variables a que caracterizan fechas especiales.  Se carga un CSV y se realiza un left join a la tabla Total_Dataset_Freq


```{r}
Dias_Especiales <- read.csv(file="../data/Caracterizacion_v2.csv", encoding="UTF-8", header=TRUE, sep=";",stringsAsFactors=FALSE)
```

```{r}
head(Dias_Especiales)
```

Se transforma la variable FEcha a formato Fecha pues se cargo como factor y se unifica el nombre a FECHA para que coincida con l clave de la tabla Total_Dataset_Freq

```{r}
Dias_Especiales <- Dias_Especiales[1:2922,]
Dias_Especiales$Fecha <- as.Date(Dias_Especiales$Fecha, format="%d-%m-%Y")
```




Estos dias especiales se incorporan al dataframe Total_Dataset_Freq
```{r}
Total_Dataset_Freq <- sqldf("SELECT * 
              FROM Total_Dataset_Freq TDF
              LEFT JOIN Dias_Especiales DE ON (TDF.FECHA=DE.Fecha)")
```

```{r}
Total_Dataset_Freq$Fecha<- NULL
```


```{r}
head(Total_Dataset_Freq)
```

```{r Frecuencia_Diaria}
save(Total_Dataset_Freq,file="../data/Frecuencia_De_Accidentes_Diario.Rda")
```

## FRECUENCIA MENSUAL

Agrupamiento de las variables explicativas en la periodicidad Mensual


```{r}
Dias_Especiales$ANO <- as.factor(format(Dias_Especiales$Fecha,'%Y'))
```

Se agrega la variable semana a la caracterizacion de los dias especiales

```{r}
Dias_Especiales$SEMANA <-as.factor(format(Dias_Especiales$Fecha,'%V'))
```

Se agrega la variable semana a la caracterizacion de los dias especiales

```{r}
Dias_Especiales$MES <-as.factor(format(Dias_Especiales$Fecha,'%m'))
```

```{r}
head(Dias_Especiales)
```


```{r Frecuencia_Mensual}
save(Dias_Especiales,file="../data/Dias_Especiales_Diario.Rda")
```



```{r}
library(sqldf)
Dias_Especiales_Mensuales <- sqldf("SELECT ANO,
CASE WHEN ANO<2016 THEN 0 ELSE 1
END Ano_Base, MES, sum(Lunes)-4 AS Freq_lunes, sum(martes)-4 AS Freq_martes, sum(miercoles)-4 AS Freq_miercoles, sum(jueves)-4 AS Freq_jueves, sum(viernes)-4 AS Freq_viernes, sum(sabado)-4 AS Freq_sabado, sum(domingo)-4 AS Freq_domingo, sum(Feriado_Lunes+Feriado_Otro) as Feriados, sum(Feriado_Lunes) AS Feriados_Lunes, sum(Feriado_Otro) AS Feriados_Otro
       FROM Dias_Especiales
       GROUP BY ANO, MES")
```

```{r Frecuencia_Mensual}
save(Dias_Especiales_Mensuales,file="../data/Dias_Especiales_Mensuales.Rda")
```


Agrupamientos de los accidentes de transitos mensual con cruce de variables descriptoras

```{r}
library(sqldf)
Total_Dataset_Freq_M <- sqldf("

SELECT TDF.ANO, 
DEM.Ano_Base, TDF.MES, TDF.GRAVEDAD, TDF.FREQ, Freq_lunes, Freq_martes, Freq_miercoles, Freq_jueves,Freq_viernes, Freq_sabado, Freq_domingo, Feriados, Feriados_Lunes, Feriados_Otro
FROM
(
SELECT  
ANO, MES, GRAVEDAD, sum(FREQ) as FREQ
FROM Total_Dataset_Freq
GROUP BY 1,2,3) TDF
        LEFT JOIN Dias_Especiales_Mensuales DEM
        ON (TDF.ANO=DEM.ANO AND TDF.MES=DEM.MES)
       GROUP BY TDF.ANO, TDF.MES, TDF.GRAVEDAD")
```


```{r Frecuencia_Mensual}
save(Total_Dataset_Freq_M,file="../data/Frecuencia_De_Accidentes_Mensual.Rda")
```


## FRECUENCIA SEMANAL

Agrupamiento de las variables explicativas en la periodicidad Semanal



```{r}
library(sqldf)
Dias_Especiales_Semanal <- sqldf("SELECT ANO, 
CASE WHEN ANO<2016 THEN 0 ELSE 1
END Ano_Base,
SEMANA, Feria_Flores_Semana,Semana_Santa_Semana, sum(Feriado_Lunes) as Feriados_Lunes, sum(Feriado_Otro) as Feriados_Otros
       FROM Dias_Especiales
       GROUP BY ANO,SEMANA,Feria_Flores_Semana,Semana_Santa_Semana")
```

```{r}
save(Dias_Especiales_Semanal,file="../data/Dias_Especiales_Semanal.Rda")
```

```{r}
library(sqldf)
Total_Dataset_Freq_S <- sqldf("

SELECT TDF.ANO, DES.Ano_Base, TDF.SEMANA, TDF.GRAVEDAD, TDF.FREQ, DES.Feria_Flores_Semana AS Feria_Flores, DES.Semana_Santa_Semana AS Semana_Santa, DES.Feriados_Lunes, DES.Feriados_Otros
FROM
(
SELECT  
ANO, SEMANA, GRAVEDAD, sum(FREQ) as FREQ
FROM Total_Dataset_Freq
GROUP BY 1,2,3) TDF
        LEFT JOIN Dias_Especiales_Semanal DES
        ON (TDF.ANO=DES.ANO AND TDF.SEMANA=DES.SEMANA)
       GROUP BY TDF.ANO, TDF.SEMANA, TDF.GRAVEDAD")
```


```{r}
save(Total_Dataset_Freq_S,file="../data/Frecuencia_De_Accidentes_Semanal.Rda")
```
