==> Entrega/Procesamiento/3-Modelo_Freq_Diaria.Rmd <==
---
title: "MODELO FRECUENCIA DIARIA"
author: "Edwin Villarraga-Jorge Montoya-Juan David Arango"
date: "10/8/2019"
output:
    html_document:
      theme: yeti
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sqldf)
```


# EVALUACIÓN DE MODELOS

## 1. Importación y estandarización de la información

```{r}
load("../data/Frecuencia_De_Accidentes_Diario.Rda")
```

```{r}
load("../data/Dias_Especiales_Diario.Rda")
```


Se crean las columnas de accidentes Graves y leves para saber la frecuencia por día

```{r}
library(reshape)
Total_Dataset_Freq <- cast(Total_Dataset_Freq[,c(1,2,3)],FECHA~GRAVEDAD)
```
Se agrega la columna TOTAL_ACCIDENTES

```{r}
Total_Dataset_Freq$TOTAL_ACCIDENTES <- Total_Dataset_Freq$ACCIDENTES_GRAVES + Total_Dataset_Freq$ACCIDENTES_LEVES
```

```{r}
Total_Dataset_Freq <- sqldf("SELECT * 
              FROM Total_Dataset_Freq
              LEFT JOIN Dias_Especiales USING(FECHA)")
```

```{r}
Total_Dataset_Freq$DIA <-as.factor(format(Total_Dataset_Freq$FECHA,'%u'))
```

```{r}
save(Total_Dataset_Freq,file="../Modelos/Total_Dataset_Freq_diaria.Rda")
```

## 2. Partición de los datos para entrenamiento y test

Se ajustarán modelos con la información disponible desde el 01 de enero de 2014 hasta el 31 de diciembre de 2017 y se utilizará el año 2018 para validar el modelo:


```{r}
Train_D_Dataset <- subset(Total_Dataset_Freq, ANO!="2018")
```

```{r}
summary(Train_D_Dataset$ANO)
```

Se ajustan otra vez los niveles del factor ANO

```{r}
Train_D_Dataset$ANO <- factor(Train_D_Dataset$ANO)
```

```{r}
summary(Train_D_Dataset$ANO)
```




```{r}
library(sqldf)
Test_D_Dataset <- sqldf("SELECT *  
       FROM Total_Dataset_Freq
       WHERE ANO == 2018")
```

```{r}
summary(Test_D_Dataset$ANO)
```

Se ajustan otra vez los niveles del factor ANO

```{r}
Test_D_Dataset$ANO <- factor(Test_D_Dataset$ANO)
```

```{r}
summary(Test_D_Dataset$ANO)
```

## 3. Selección de las mejores variables para el modelo 

Se utilizará el método forward selection para elegir las mejores variables explicativas del modelo teniendo como criterio aquellas variables que presente mejor R^2 ajustado

#### Selección de variables para Total accidentes
```{r}
library (leaps)
regfit.fwd=regsubsets (TOTAL_ACCIDENTES∼ANO+MES+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Previo_feriado+Prima+Mujer+Padre+Madre+AmoryAmistad+Semana_Santa+Viernes_Antes_Puente+Quincena+Viernes_Desp_Quincena_v1+Viernes_Desp_Quincena_v2+Feria_Flores,Train_D_Dataset, method ="forward", nvmax= 80)
summary (regfit.fwd)


```

```{r}
reg.summary =summary(regfit.fwd)
names(reg.summary)
```
```{r}
reg.summary$rsq
```

Selección de variables con el mejor R^2 ajustado
```{r}
max_adjr<-which.max (reg.summary$adjr2)
max_adjr
```


```{r}
par(mfrow =c(2,2))
plot(reg.summary$rss ,xlab=" Number of Variables ",ylab=" RSS",
type="l")
plot(reg.summary$adjr2 ,xlab =" Number of Variables ",
ylab=" Adjusted RSq",type="l")
points (max_adjr, reg.summary$adjr2[max_adjr], col ="red",cex =2, pch =20)
```

```{r}
plot(regfit.fwd ,scale ="adjr2")
```

```{r}
coef(regfit.fwd ,max_adjr)
```



## 4. Modelamiento

```{r}
set.seed(123) # fija la semilla del generador de parámetros para que sea reproducible
```


Se realiza el modelo de regresión lineal con las variables seleccionadas y se revisa el p-valor de cada una para seleccionar las variables definitivas del modelo 
```{r}
library(caret)
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit = caret::train(TOTAL_ACCIDENTES∼ANO+MES+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Previo_feriado+Mujer+Padre+Madre+AmoryAmistad+Semana_Santa+Viernes_Antes_Puente+Quincena+Viernes_Desp_Quincena_v1+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
summary(caret_lm_fit)
```


### 1. REGRESION LINEAL

#### Total Accidentes
```{r}
head(Train_D_Dataset)
```

```{r}
library(caret)
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit = caret::train(TOTAL_ACCIDENTES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
summary(caret_lm_fit)
```


```{r}
caret_lm_fit
```

Calculo MSE y RMSE para los datos de entrenamiento

Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores


```{r}
y_tr_pred_lm<-predict(caret_lm_fit,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])

```

```{r}
mse_tr_lm<-mean((Train_D_Dataset$TOTAL_ACCIDENTES-y_tr_pred_lm)^2) # calcula el mse de entrenamiento
RMSE_tr_lm = sqrt(mse_tr_lm)
mse_tr_lm
RMSE_tr_lm
```

Calculo MSE y RMSE para los datos de validación

```{r}
y_test_pred_lm<-predict(caret_lm_fit,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_lm<-mean((Test_D_Dataset$TOTAL_ACCIDENTES-y_test_pred_lm)^2) # calcula el mse de entrenamiento
RMSE_test_lm = sqrt(mse_test_lm)
mse_test_lm
RMSE_test_lm
```

Predicción en la muestra

```{r}
library(plotly)
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_lm,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```


Gráfica serie 2018

```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_lm,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```

#### Accidentes Graves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_m = caret::train(ACCIDENTES_GRAVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_lm_fit_m)
```

```{r}
caret_lm_fit_m
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_lm_m<-predict(caret_lm_fit_m,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])

```

```{r}
mse_tr_lm_m<-mean((Train_D_Dataset$ACCIDENTES_GRAVES-y_tr_pred_lm_m)^2) # calcula el mse de entrenamiento
RMSE_tr_lm_m = sqrt(mse_tr_lm_m)
mse_tr_lm_m
RMSE_tr_lm_m
```


Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_lm_m<-predict(caret_lm_fit_m,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_lm_m<-mean((Test_D_Dataset$ACCIDENTES_GRAVES-y_test_pred_lm_m)^2) # calcula el mse de entrenamiento
RMSE_test_lm_m = sqrt(mse_test_lm_m)
mse_test_lm_m
RMSE_test_lm_m
```

Predicción en la muestra 

```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_lm_m,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_lm_m,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))
```



#### Accidentes Leves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_lm_fit_sd)
```

```{r}
caret_lm_fit_sd
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_lm_sd<-predict(caret_lm_fit_sd,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])

```

```{r}
mse_tr_lm_sd<-mean((Train_D_Dataset$ACCIDENTES_LEVES-y_tr_pred_lm_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_lm_sd = sqrt(mse_tr_lm_sd)
mse_tr_lm_sd
RMSE_tr_lm_sd
```


Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_lm_sd<-predict(caret_lm_fit_sd,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_lm_sd<-mean((Test_D_Dataset$ACCIDENTES_LEVES-y_test_pred_lm_sd)^2) # calcula el mse de entrenamiento
RMSE_test_lm_sd = sqrt(mse_test_lm_sd)
mse_test_lm_sd
RMSE_test_lm_sd
```

Predicción en la muestra 

```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_lm_sd,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_lm_sd,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))
```

#### Resumen Modelos Regresión lineal para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes graves","Accidentes leves")
RMSE_Train_lm = round(c(RMSE_tr_lm,RMSE_tr_lm_m,RMSE_tr_lm_sd), 3)
RMSE_Test_lm = round(c(RMSE_test_lm,RMSE_test_lm_m,RMSE_test_lm_sd),3)

Tabla_lm = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_lm,RMSE_Test_lm))
Tabla_lm
```



### 2. KNN





```{r}
head(Train_D_Dataset)
```

#### Total Accidentes

```{r}
library(caret)
```

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_knn_fit = caret::train(TOTAL_ACCIDENTES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "knn", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
summary(caret_knn_fit)
```

```{r}
caret_knn_fit
```

Calculo MSE y RMSE para los datos de entrenamiento

```{r}
y_tr_pred_knn<-predict(caret_knn_fit,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_knn<-mean((Train_D_Dataset$TOTAL_ACCIDENTES-y_tr_pred_knn)^2) # calcula el mse de entrenamiento
RMSE_tr_knn = sqrt(mse_tr_knn)
mse_tr_knn
RMSE_tr_knn
```

Calculo MSE y RMSE para los datos de validación 

```{r}
y_test_pred_knn<-predict(caret_knn_fit,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_knn<-mean((Test_D_Dataset$TOTAL_ACCIDENTES-y_test_pred_knn)^2) # calcula el mse de entrenamiento
RMSE_test_knn = sqrt(mse_test_knn)
mse_test_knn
RMSE_test_knn
```
Predicción en la muestra 

```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_knn,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_knn,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```

#### Accidentes graves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_knn_fit_m = caret::train(ACCIDENTES_GRAVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "knn", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_knn_fit_m)
```

```{r}
caret_knn_fit_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_knn_m<-predict(caret_knn_fit_m,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_knn_m<-mean((Train_D_Dataset$ACCIDENTES_GRAVES-y_tr_pred_knn_m)^2) # calcula el mse de entrenamiento
RMSE_tr_knn_m = sqrt(mse_tr_knn_m)
mse_tr_knn_m
RMSE_tr_knn_m
```
Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_knn_m<-predict(caret_knn_fit_m,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_knn_m<-mean((Test_D_Dataset$ACCIDENTES_GRAVES-y_test_pred_knn_m)^2) # calcula el mse de entrenamiento
RMSE_test_knn_m = sqrt(mse_test_knn_m)
mse_test_knn_m
RMSE_test_knn_m
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_knn_m,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_knn_m,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))
```


#### Accidentes Leves


```{r}
trcntrl = trainControl(method="cv", number=10)
caret_knn_fit_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "knn", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_knn_fit_sd)
```

```{r}
caret_knn_fit_sd
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_knn_sd<-predict(caret_knn_fit_sd,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_knn_sd<-mean((Train_D_Dataset$ACCIDENTES_LEVES-y_tr_pred_knn_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_knn_sd = sqrt(mse_tr_knn_sd)
mse_tr_knn_sd
RMSE_tr_knn_sd
```
Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_knn_sd<-predict(caret_knn_fit_sd,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_knn_sd<-mean((Test_D_Dataset$ACCIDENTES_LEVES-y_test_pred_knn_sd)^2) # calcula el mse de entrenamiento
RMSE_test_knn_sd = sqrt(mse_test_knn_sd)
mse_test_knn_sd
RMSE_test_knn_sd
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_knn_sd,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_knn_sd,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```


#### Resumen Modelos KNN para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes graves","Accidentes leves")
RMSE_Train_knn = round(c(RMSE_tr_knn,RMSE_tr_knn_m,RMSE_tr_knn_sd), 3)
RMSE_Test_knn = round(c(RMSE_test_knn,RMSE_test_knn_m,RMSE_test_knn_sd),3)

Tabla_knn = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_knn,RMSE_Test_knn))
Tabla_knn
```


### 3. MODELO LINEAL GENERALIZADO


#### Total Accidentes

```{r}
glm_fit<-glm(TOTAL_ACCIDENTES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset, family = "poisson")
summary(glm_fit)
```

```{r}
glm_fit
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_glm<-predict(glm_fit,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")],type="response")
mse_tr_glm<-mean((Train_D_Dataset$TOTAL_ACCIDENTES-y_tr_pred_glm)^2) # calcula el mse de entrenamiento
RMSE_tr_glm = sqrt(mse_tr_glm)
mse_tr_glm
RMSE_tr_glm
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_glm<-predict(glm_fit,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")],type="response")
mse_test_glm<-mean((Train_D_Dataset$TOTAL_ACCIDENTES-y_test_pred_glm)^2) # calcula el mse de entrenamiento
```
```{r}
RMSE_test_glm = sqrt(mse_test_glm)
mse_test_glm
RMSE_test_glm
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_glm,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_glm,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```


#### Accidentes Graves

```{r}
glm_fit_m<-glm(ACCIDENTES_GRAVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset, family = "poisson")
summary(glm_fit)
```

```{r}
glm_fit_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_glm_m<-predict(glm_fit_m,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")],type="response")
mse_tr_glm_m<-mean((Train_D_Dataset$ACCIDENTES_GRAVES-y_tr_pred_glm_m)^2) # calcula el mse de entrenamiento
RMSE_tr_glm_m = sqrt(mse_tr_glm_m)
mse_tr_glm_m
RMSE_tr_glm_m
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_glm_m<-predict(glm_fit_m,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")],type="response")
mse_test_glm_m<-mean((Train_D_Dataset$ACCIDENTES_GRAVES-y_test_pred_glm_m)^2) # calcula el mse de entrenamiento
```
```{r}
RMSE_test_glm_m = sqrt(mse_test_glm_m)
mse_test_glm_m
RMSE_test_glm_m
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_glm_m,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_glm_m,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))
```



#### Accidentes Leves

```{r}
glm_fit_sd<-glm(ACCIDENTES_LEVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset, family = "poisson")
summary(glm_fit_sd)
```

```{r}
glm_fit_sd
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_glm_sd<-predict(glm_fit_sd,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")],type="response")
mse_tr_glm_sd<-mean((Train_D_Dataset$ACCIDENTES_LEVES-y_tr_pred_glm_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_glm_sd = sqrt(mse_tr_glm_sd)
mse_tr_glm_sd
RMSE_tr_glm_sd
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_glm_sd<-predict(glm_fit_sd,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")],type="response")
mse_test_glm_sd<-mean((Train_D_Dataset$ACCIDENTES_LEVES-y_test_pred_glm_sd)^2) # calcula el mse de entrenamiento
```
```{r}
RMSE_test_glm_sd = sqrt(mse_test_glm_sd)
mse_test_glm_sd
RMSE_test_glm_sd
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_glm_sd,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))
```

Gráfica serie 2018  
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_glm_sd,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))
```
#### REsumen Modelos Regresión lineal generalizado para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes Graves","Accidentes Leves")
RMSE_Train_glm = round(c(RMSE_tr_glm,RMSE_tr_glm_m,RMSE_tr_glm_sd), 3)
RMSE_Test_glm = round(c(RMSE_test_glm,RMSE_test_glm_m,RMSE_test_glm_sd),3)

Tabla_glm = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_glm,RMSE_Test_glm))
Tabla_glm
```




### 4. ARBOLES DE REGRESION

#### Total Accidentes

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_tree_fit = caret::train(TOTAL_ACCIDENTES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores,data=Train_D_Dataset,
                              method = "rpart", trControl = trcntrl,
                      parms = list(split = "gini"),
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```



```{r}
caret_tree_fit
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_tree<-predict(caret_tree_fit,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_tree<-mean((Train_D_Dataset$TOTAL_ACCIDENTES-y_tr_pred_tree)^2) # calcula el mse de entrenamiento
RMSE_tr_tree = sqrt(mse_tr_tree)
mse_tr_tree
RMSE_tr_tree
```
Calculo MSE y RMSE para los datos de validación 
```{r}
  y_test_pred_tree<-predict(caret_tree_fit,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_tree<-mean((Test_D_Dataset$TOTAL_ACCIDENTES-y_test_pred_tree)^2) # calcula el mse de entrenamiento
RMSE_test_tree = sqrt(mse_test_tree)
mse_test_tree
RMSE_test_tree
```
Predicción en la muestra  
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_tree,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_tree,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```


#### Accidentes graves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_tree_fit_m = caret::train(ACCIDENTES_GRAVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores,data=Train_D_Dataset,
                              method = "rpart", trControl = trcntrl,
                      parms = list(split = "gini"),
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
caret_tree_fit_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_tree_m<-predict(caret_tree_fit_m,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_tree_m<-mean((Train_D_Dataset$ACCIDENTES_GRAVES-y_tr_pred_tree_m)^2) # calcula el mse de entrenamiento
RMSE_tr_tree_m = sqrt(mse_tr_tree_m)
mse_tr_tree_m
RMSE_tr_tree_m
```
Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_tree_m<-predict(caret_tree_fit_m,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_tree_m<-mean((Test_D_Dataset$ACCIDENTES_GRAVES-y_test_pred_tree_m)^2) # calcula el mse de entrenamiento
RMSE_test_tree_m = sqrt(mse_test_tree_m)
mse_test_tree_m
RMSE_test_tree_m
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_tree_m,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_tree_m,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```



#### Accidentes leves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_tree_fit_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores,data=Train_D_Dataset,
                              method = "rpart", trControl = trcntrl,
                      parms = list(split = "gini"),
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
caret_tree_fit_sd
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_tree_sd<-predict(caret_tree_fit_sd,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_tree_sd<-mean((Train_D_Dataset$ACCIDENTES_LEVES-y_tr_pred_tree_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_tree_sd = sqrt(mse_tr_tree_sd)
mse_tr_tree_sd
RMSE_tr_tree_sd
```
Calculo MSE y RMSE para los datos de validación 
```{r}
  y_test_pred_tree_sd<-predict(caret_tree_fit_sd,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_tree_sd<-mean((Test_D_Dataset$ACCIDENTES_LEVES-y_test_pred_tree_sd)^2) # calcula el mse de entrenamiento
RMSE_test_tree_sd = sqrt(mse_test_tree_sd)
mse_test_tree_sd
RMSE_test_tree_sd
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_tree_sd,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_tree_sd,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))
```

#### Resumen Modelos Árboles de regresión para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes Graves","Accidentes Leves")
RMSE_Train_tree = round(c(RMSE_tr_tree,RMSE_tr_tree_m,RMSE_tr_tree_sd), 3)
RMSE_Test_tree = round(c(RMSE_test_tree,RMSE_test_tree_m,RMSE_test_tree_sd),3)

Tabla_tree = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_tree,RMSE_Test_tree))
Tabla_tree
```

### 5. BOSQUE ALEATORIO


#### Total Accidentes

```{r Random_Forest_TOTAL_ACCIDENTES}
trcntrl = trainControl(method="cv", number=10)
caret_rf_fit = caret::train(TOTAL_ACCIDENTES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "rf", trControl = trcntrl,
                      prox=TRUE,allowParallel=TRUE)
```


```{r}
summary(caret_rf_fit)
```

```{r}
caret_rf_fit
```

```{r}
plot(caret_rf_fit)
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_rf<-predict(caret_rf_fit,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_rf<-mean((Train_D_Dataset$TOTAL_ACCIDENTES-y_tr_pred_rf)^2) # calcula el mse de entrenamiento
RMSE_tr_rf = sqrt(mse_tr_rf)
mse_tr_rf
RMSE_tr_rf
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_rf<-predict(caret_rf_fit,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_rf<-mean((Test_D_Dataset$TOTAL_ACCIDENTES-y_test_pred_rf)^2) # calcula el mse de entrenamiento
RMSE_test_rf = sqrt(mse_test_rf)
mse_test_rf
RMSE_test_rf
```


Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_rf,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y=  ~y_test_pred_rf,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 0.75, y = 0.9))

```


#### Accidentes Graves

```{r Random_Forest_ACCIDENTES_GRAVES}
trcntrl = trainControl(method="cv", number=10)
caret_rf_fit_m = caret::train(ACCIDENTES_GRAVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "rf", trControl = trcntrl,
                      prox=TRUE,allowParallel=TRUE)
```


```{r}
summary(caret_rf_fit_m)
```

```{r}
caret_rf_fit_m
```

```{r}
plot(caret_rf_fit_m)
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_rf_m<-predict(caret_rf_fit_m,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_rf_m<-mean((Train_D_Dataset$ACCIDENTES_GRAVES-y_tr_pred_rf_m)^2) # calcula el mse de entrenamiento
RMSE_tr_rf_m = sqrt(mse_tr_rf_m)
mse_tr_rf_m
RMSE_tr_rf_m
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_rf_m<-predict(caret_rf_fit_m,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_rf_m<-mean((Test_D_Dataset$ACCIDENTES_GRAVES-y_test_pred_rf_m)^2) # calcula el mse de entrenamiento
RMSE_test_rf_m = sqrt(mse_test_rf_m)
mse_test_rf_m
RMSE_test_rf_m
```

Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_rf_m,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_rf_m,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 0.75, y = 0.9))

```

#### Accidentes Leves

```{r Random_Forest_ACCIDENTES_LEVES}
trcntrl = trainControl(method="cv", number=10)
caret_rf_fit_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Train_D_Dataset,
                      method = "rf", trControl = trcntrl,
                      prox=TRUE,allowParallel=TRUE)
```


```{r}
summary(caret_rf_fit_sd)
```

```{r}
caret_rf_fit_sd
```

```{r}
plot(caret_rf_fit_sd)
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_rf_sd<-predict(caret_rf_fit_sd,Train_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_tr_rf_sd<-mean((Train_D_Dataset$ACCIDENTES_LEVES-y_tr_pred_rf_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_rf_sd = sqrt(mse_tr_rf_sd)
mse_tr_rf_sd
RMSE_tr_rf_sd
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_rf_sd<-predict(caret_rf_fit_sd,Test_D_Dataset[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
mse_test_rf_sd<-mean((Test_D_Dataset$ACCIDENTES_LEVES-y_test_pred_rf_sd)^2) # calcula el mse de entrenamiento
RMSE_test_rf_sd = sqrt(mse_test_rf_sd)
mse_test_rf_sd
RMSE_test_rf_sd
```


Predicción en la muestra 
```{r}
plot_ly (data=Train_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_rf_sd,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_D_Dataset,
         x = ~FECHA,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_rf_sd,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 0.75, y = 0.9))

```


#### Resumen Modelos Random Forest para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Total Graves","Total Leves")
RMSE_Train_rf = round(c(RMSE_tr_rf,RMSE_tr_rf_m,RMSE_tr_rf_sd), 3)
RMSE_Test_rf = round(c(RMSE_test_rf,RMSE_test_rf_m,RMSE_test_rf_sd),3)

Tabla_rf = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_rf,RMSE_Test_rf))
Tabla_rf
```



## ELECCION DEL MODELO


### 1. Elección del modelo para el total de accidentes

Comparación en el entrenamiento
```{r}
comparacion_tr<-data.frame(FECHA=Total_Dataset_Freq$FECHA[Total_Dataset_Freq$FECHA<="2017-12-31"],
                          ACCIDENTES=Total_Dataset_Freq$TOTAL_ACCIDENTES[Total_Dataset_Freq$FECHA<="2017-12-31"],
                           lm= y_tr_pred_lm, 
                           knn= y_tr_pred_knn, 
                           glm=y_tr_pred_glm ,
                           arbol=y_tr_pred_tree,
                           rf=y_tr_pred_rf)
```

```{r}
plot_ly (data=comparacion_tr,
         x = ~FECHA,
         y = ~ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
    add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Total accidentes (Entrenamiento)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```


Comparación en la validación
```{r}
comparacion_vl<-data.frame(FECHA=Test_D_Dataset$FECHA,
                          ACCIDENTES=Test_D_Dataset$TOTAL_ACCIDENTES,
                           lm= y_test_pred_lm, 
                           knn= y_test_pred_knn, 
                           glm=y_test_pred_glm ,
                           arbol=y_test_pred_tree,
                           rf=y_test_pred_rf)
```

```{r}
plot_ly (data=comparacion_vl,
         x = ~FECHA,
         y = ~ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
  add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Total Accidentes (Validación)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```




#### Comparación con los RMSE:

```{r}
Entrenamiento<-round(c(RMSE_tr_lm,RMSE_tr_knn,RMSE_tr_glm,RMSE_tr_tree,RMSE_tr_rf),3) 
Validacion<-round(c(RMSE_test_lm,RMSE_test_knn,RMSE_test_glm,RMSE_test_tree,RMSE_test_rf),3) 
nombres<-c("lm","knn","glm","árbol","bosque")
ResultadosRMSE<-data.frame(Entrenamiento=Entrenamiento,Validacion=Validacion)
rownames(ResultadosRMSE)<-nombres
```

Cálculo de la variación

```{r}
ResultadosRMSE$Por_variacion<-((ResultadosRMSE$Validacion-ResultadosRMSE$Entrenamiento)/ResultadosRMSE$Entrenamiento)*100
ResultadosRMSE
```


### 2. Elección del modelo para Accidentes graves

Comparación en el entrenamiento
```{r}
comparacion_tr<-data.frame(FECHA=Total_Dataset_Freq$FECHA[Total_Dataset_Freq$FECHA<="2017-12-31"],
                          ACCIDENTESG=Total_Dataset_Freq$ACCIDENTES_GRAVES[Total_Dataset_Freq$FECHA<="2017-12-31"],
                           lm= y_tr_pred_lm_m, 
                           knn= y_tr_pred_knn_m, 
                           glm=y_tr_pred_glm_m,
                           arbol=y_tr_pred_tree_m,
                           rf=y_tr_pred_rf_m)
```

```{r}
plot_ly (data=comparacion_tr,
         x = ~FECHA,
         y = ~ACCIDENTESG,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbo',
            line=list(width=1,color="green"))%>%
    add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes graves (Entrenamiento)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```


Comparación en la validación
```{r}
comparacion_vl<-data.frame(FECHA=Test_D_Dataset$FECHA,
                          ACCIDENTESG=Test_D_Dataset$ACCIDENTES_GRAVES,
                           lm= y_test_pred_lm_m, 
                           knn= y_test_pred_knn_m, 
                           glm=y_test_pred_glm_m,
                           arbol=y_test_pred_tree_m,
                           rf=y_test_pred_rf_m)
```

```{r}
plot_ly (data=comparacion_vl,
         x = ~FECHA,
         y = ~ACCIDENTESG,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
  add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes graves (Validación)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```



#### Comparación con los RMSE:

```{r}
Entrenamiento<-round(c(RMSE_tr_lm_m,RMSE_tr_knn_m,RMSE_tr_glm_m,RMSE_tr_tree_m,RMSE_tr_rf_m),3) 
Validacion<-round(c(RMSE_test_lm_m,RMSE_test_knn_m,RMSE_test_glm_m,RMSE_test_tree_m,RMSE_test_rf_m),3) 
nombres<-c("lm","knn","glm","árbol","bosque")
ResultadosRMSE<-data.frame(Entrenamiento=Entrenamiento,Validacion=Validacion)
rownames(ResultadosRMSE)<-nombres
```

Cálculo de la variación

```{r}
ResultadosRMSE$Por_variacion<-((ResultadosRMSE$Validacion-ResultadosRMSE$Entrenamiento)/ResultadosRMSE$Entrenamiento)*100
ResultadosRMSE
```

### 3. Elección del modelo para Accidentes leves

Comparación en el entrenamiento
```{r}
comparacion_tr<-data.frame(FECHA=Total_Dataset_Freq$FECHA[Total_Dataset_Freq$FECHA<="2017-12-31"],
                          ACCIDENTESL=Total_Dataset_Freq$ACCIDENTES_LEVES[Total_Dataset_Freq$FECHA<="2017-12-31"],
                           lm= y_tr_pred_lm_sd, 
                           knn= y_tr_pred_knn_sd, 
                           glm=y_tr_pred_glm_sd,
                           arbol=y_tr_pred_tree_sd,
                           rf=y_tr_pred_rf_sd)
```

```{r}
plot_ly (data=comparacion_tr,
         x = ~FECHA,
         y = ~ACCIDENTESL,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbo',
            line=list(width=1,color="green"))%>%
    add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes leves (Entrenamiento)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```


Comparación en la validación
```{r}
comparacion_vl<-data.frame(FECHA=Test_D_Dataset$FECHA,
                          ACCIDENTESL=Test_D_Dataset$ACCIDENTES_GRAVES,
                           lm= y_test_pred_lm_sd, 
                           knn= y_test_pred_knn_sd, 
                           glm=y_test_pred_glm_sd,
                           arbol=y_test_pred_tree_sd,
                           rf=y_test_pred_rf_sd)
```

```{r}
plot_ly (data=comparacion_vl,
         x = ~FECHA,
         y = ~ACCIDENTESL,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
  add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes leves (Validación)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```

#### Comparación con los RMSE:

```{r}
Entrenamiento<-round(c(RMSE_tr_lm_sd,RMSE_tr_knn_sd,RMSE_tr_glm_sd,RMSE_tr_tree_sd,RMSE_tr_rf_sd),3) 
Validacion<-round(c(RMSE_test_lm_sd,RMSE_test_knn_sd,RMSE_test_glm_sd,RMSE_test_tree_sd,RMSE_test_rf_sd),3)
nombres<-c("lm","knn","glm","árbol","bosque")
ResultadosRMSE<-data.frame(Entrenamiento=Entrenamiento,Validacion=Validacion)
rownames(ResultadosRMSE)<-nombres
```

Cálculo de la variación

```{r}
ResultadosRMSE$Por_variacion<-((ResultadosRMSE$Validacion-ResultadosRMSE$Entrenamiento)/ResultadosRMSE$Entrenamiento)*100
ResultadosRMSE
```

#### Modelos elegidos

Teniendo como criterio el mínimo RMSE y que la variación entre los datos de entrenamiento y validación no superen el 15 %, se eligieron los siguientes modelos: 

##### Modelo de regresión lineal para predición de Total Accidentes 

Se ajusta el modelo con todos los datos desde el 01-01-2014 al 31-12-2018

```{r}
library(caret)
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_final = caret::train(TOTAL_ACCIDENTES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Total_Dataset_Freq,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
summary(caret_lm_fit_final)
```


Se guardan el modelo en un objeto de r
```{r}
saveRDS(caret_lm_fit_final,"../Modelos/Prediccion_Total_Diario.rds")
```


```{r}
Modelo_Total_diario<-readRDS(file="../Modelos/Prediccion_Total_Diario.rds")
```


##### Modelo de árbol de regresión para predición de Accidentes Graves

Se ajusta el modelo con todos los datos desde el 01-01-2014 al 31-12-2018
```{r}
trcntrl = trainControl(method="cv", number=10)
caret_tree_fit_m_final = caret::train(ACCIDENTES_GRAVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores,data=Total_Dataset_Freq,
                              method = "rpart", trControl = trcntrl,
                      parms = list(split = "gini"),
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
caret_tree_fit_m_final
```


Se guardan el modelo en un objeto de r
```{r}
saveRDS(caret_tree_fit_m_final,"../Modelos/Prediccion_Grave_Diario.rds")
```


```{r}
Modelo_Grave_diario<-readRDS(file="../Modelos/Prediccion_Grave_Diario.rds")
```


##### Modelo de regresión lineal para predición de Accidentes Leves

Se ajusta el modelo con todos los datos desde el 01-01-2014 al 31-12-2018
```{r}
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_sd_final = caret::train(ACCIDENTES_LEVES∼Ano_Base+DIA+SEMANA+Feriado_Lunes+Feriado_Otro+Madre+Semana_Santa+Viernes_Desp_Quincena_v2+Feria_Flores, data=Total_Dataset_Freq,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_lm_fit_sd_final)
```



Se guardan el modelo en un objeto de r
```{r}
saveRDS(caret_lm_fit_sd_final,"../Modelos/Prediccion_leves_Diario.rds")
```


```{r}
Modelo_leves_diario<-readRDS(file="../Modelos/Prediccion_leves_Diario.rds")
```

##### Datos para pronóstico

Se oganizan los datos necesarios para el pronóstico de los accidentes en los años 2019, 2020 y 2021

Importación de los datos
```{r}
load("../data/Dias_Especiales_Diario.Rda")
```

```{r}
Dias_Especiales$DIA <-as.factor(format(Dias_Especiales$Fecha,'%u'))
head(Dias_Especiales)
```

```{r}
datos_pronostico_diario<-Dias_Especiales[,c("Fecha","Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")]
```


Predicción del Total de accidentes con el modelo de regresión lineal

```{r}
datos_pronostico_diario$prediccion_Total<-predict(Modelo_Total_diario,datos_pronostico_diario[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
```

Predicción de accidentes graves con el modelo de árbol de regresión

```{r}
datos_pronostico_diario$prediccion_Graves<-predict(Modelo_Grave_diario,datos_pronostico_diario[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
```


Predicción de accidentes leves con el modelo de regresión lineal

```{r}
datos_pronostico_diario$prediccion_Leves<-predict(Modelo_leves_diario,datos_pronostico_diario[,c("Ano_Base","DIA","SEMANA","Feriado_Lunes","Feriado_Otro","Madre","Semana_Santa","Viernes_Desp_Quincena_v2","Feria_Flores")])
```

Se guardan los datos de pronóstico en un objeto de r
```{r}
save(datos_pronostico_diario,file="../Modelos/datos_pronostico_diario.Rda")
```
