==> Entrega/Procesamiento/5-Modelo_Freq_Mensual.Rmd <==
---
title: "MODELO FRECUENCIA MENSUAL"
author: "Edwin Villarraga-Jorge Montoya-Juan David Arango"
date: "10/8/2019"
output:
    html_document:
      theme: yeti
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sqldf)
```


# EVALUACIÓN DE MODELOS

## 1. Importación y estandarización de la información

```{r}
load("../data/Frecuencia_De_Accidentes_Mensual.Rda")
```

```{r}
load("../data/Dias_Especiales_Mensuales.Rda")
```




Se crean las columnas de accidentes Graves y leves para saber la frecuencia por día

```{r}
library(reshape)
Total_Dataset_Freq_M <- cast(Total_Dataset_Freq_M[,c(1,3,4,5)],ANO+MES~GRAVEDAD)
```
Se agrega la columna TOTAL_ACCIDENTES

```{r}
Total_Dataset_Freq_M$TOTAL_ACCIDENTES <- Total_Dataset_Freq_M$ACCIDENTES_GRAVES + Total_Dataset_Freq_M$ACCIDENTES_LEVES
```

```{r}
Total_Dataset_Freq_M <- sqldf("SELECT* 
              FROM Total_Dataset_Freq_M FS
              LEFT JOIN Dias_Especiales_Mensuales DES 
              ON (FS.ANO=DES.ANO AND FS.MES=DES.MES)")
```


```{r}
library(dplyr)
Total_Dataset_Freq_M<-unite_(Total_Dataset_Freq_M, "Ano_mes", c("ANO..6","MES..8"))
```


```{r}
save(Total_Dataset_Freq_M,file="../Modelos/Total_Dataset_Freq_M_mensual.Rda")
```

## 2. Partición de los datos para entrenamiento y test

Se ajustarán modelos con la información disponible desde el 01 de enero de 2014 hasta el 31 de diciembre de 2017 y se utilizará el año 2018 para validar el modelo:


```{r}
Train_M_Dataset <- subset(Total_Dataset_Freq_M, ANO!="2018")
```

```{r}
summary(Train_M_Dataset $ANO)
```

Se ajustan otra vez los niveles del factor ANO

```{r}
Train_M_Dataset $ANO <- factor(Train_M_Dataset $ANO)
```

```{r}
summary(Train_M_Dataset $ANO)
```




```{r}
library(sqldf)
Test_M_Dataset <- sqldf("SELECT *  
       FROM Total_Dataset_Freq_M
       WHERE ANO == 2018")
```

```{r}
summary(Test_M_Dataset$ANO)
```

Se ajustan otra vez los niveles del factor ANO

```{r}
Test_M_Dataset$ANO <- factor(Test_M_Dataset$ANO)
```

```{r}
summary(Test_M_Dataset$ANO)
```

## 3. Selección de las mejores variables para el modelo 

Se utilizará el método forward selection para elegir las mejores variables explicativas del modelo teniendo como criterio aquellas variables que presente mejor R^2 ajustado



```{r}
library (leaps)
regfit.fwd=regsubsets (TOTAL_ACCIDENTES∼Ano_Base+MES+Freq_lunes+Freq_martes+Freq_miercoles+Freq_jueves+Freq_viernes+Freq_sabado+Freq_domingo+Feriados+Feriados_Lunes+Feriados_Otro,Train_M_Dataset , method ="forward", nvmax= 80)
summary (regfit.fwd)


```

```{r}
reg.summary =summary(regfit.fwd)
names(reg.summary)
```
```{r}
reg.summary$rsq
```

Selección de variables con el mejor R^2 ajustado
```{r}
max_adjr<-which.max (reg.summary$adjr2)
max_adjr
```

```{r}
par(mfrow =c(2,2))
plot(reg.summary$rss ,xlab=" Número de Variables ",ylab=" RSS",
type="l")
plot(reg.summary$adjr2 ,xlab =" Número de Variables ",
ylab=" Adjusted RSq",type="l")

points (max_adjr, reg.summary$adjr2[max_adjr], col ="red",cex =2, pch =20)
```

```{r}
plot(regfit.fwd ,scale ="adjr2")
```
Variables seleccionadas
```{r}
coef(regfit.fwd ,max_adjr)
```

## 4. Modelamiento

```{r}
set.seed(123) # fija la semilla del generador de parámetros para que sea reproducible
```


Se realiza el modelo de regresión lineal con las variables seleccionadas y se revisa el p-valor de cada una para seleccionar las variables definitivas del modelo 
```{r}
library(caret)
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_m = caret::train(TOTAL_ACCIDENTES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
summary(caret_lm_fit_m)
```


### 1. REGRESION LINEAL

#### Total Accidentes
```{r}
head(Train_M_Dataset )
```

```{r}
library(caret)
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_m = caret::train(TOTAL_ACCIDENTES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
summary(caret_lm_fit_m)
```


```{r}
caret_lm_fit_m
```

Calculo MSE y RMSE para los datos de entrenamiento


```{r}
y_tr_pred_lm_m<-predict(caret_lm_fit_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])

```

```{r}
mse_tr_lm_m<-mean((Train_M_Dataset $TOTAL_ACCIDENTES-y_tr_pred_lm_m)^2) # calcula el mse de entrenamiento
RMSE_tr_lm_m = sqrt(mse_tr_lm_m)
mse_tr_lm_m
RMSE_tr_lm_m
```

Calculo MSE y RMSE para los datos de validación

```{r}
y_test_pred_lm_m<-predict(caret_lm_fit_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_lm_m<-mean((Test_M_Dataset$TOTAL_ACCIDENTES-y_test_pred_lm_m)^2) # calcula el mse de entrenamiento
RMSE_test_lm_m = sqrt(mse_test_lm_m)
mse_test_lm_m
RMSE_test_lm_m
```

Predicción en la muestra

```{r}
library(plotly)
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_lm_m,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```


Gráfica serie 2018

```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_lm_m,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```

#### Accidentes Graves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_m_m = caret::train(ACCIDENTES_GRAVES∼MES+Feriados+Freq_domingo, data=Train_M_Dataset ,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_lm_fit_m_m)
```

```{r}
caret_lm_fit_m_m
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_lm_m_m<-predict(caret_lm_fit_m_m,Train_M_Dataset [,c("Freq_domingo","MES","Feriados")])

```

```{r}
mse_tr_lm_m_m<-mean((Train_M_Dataset $ACCIDENTES_GRAVES-y_tr_pred_lm_m_m)^2) # calcula el mse de entrenamiento
RMSE_tr_lm_m_m = sqrt(mse_tr_lm_m_m)
mse_tr_lm_m_m
RMSE_tr_lm_m_m
```


Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_lm_m_m<-predict(caret_lm_fit_m_m,Test_M_Dataset[,c("Freq_domingo","MES","Feriados")])
mse_test_lm_m_m<-mean((Test_M_Dataset$ACCIDENTES_GRAVES-y_test_pred_lm_m_m)^2) # calcula el mse de entrenamiento
RMSE_test_lm_m_m = sqrt(mse_test_lm_m_m)
mse_test_lm_m_m
RMSE_test_lm_m_m
```

Predicción en la muestra 

```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_lm_m_m,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         split = ~ANO,
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_lm_m_m,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))
```



#### Accidentes Leves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_m_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_lm_fit_m_sd)
```

```{r}
caret_lm_fit_m_sd
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_lm_m_sd<-predict(caret_lm_fit_m_sd,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])

```

```{r}
mse_tr_lm_m_sd<-mean((Train_M_Dataset $ACCIDENTES_LEVES-y_tr_pred_lm_m_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_lm_m_sd = sqrt(mse_tr_lm_m_sd)
mse_tr_lm_m_sd
RMSE_tr_lm_m_sd
```


Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_lm_m_sd<-predict(caret_lm_fit_m_sd,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_lm_m_sd<-mean((Test_M_Dataset$ACCIDENTES_LEVES-y_test_pred_lm_m_sd)^2) # calcula el mse de entrenamiento
RMSE_test_lm_m_sd = sqrt(mse_test_lm_m_sd)
mse_test_lm_m_sd
RMSE_test_lm_m_sd
```

Predicción en la muestra 

```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_lm_m_sd,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_lm_m_sd,
            name='Modelo lm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))
```

#### Resumen Modelos Regresión lineal para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes graves","Accidentes leves")
RMSE_Train_lm = round(c(RMSE_tr_lm_m,RMSE_tr_lm_m_m,RMSE_tr_lm_m_sd), 3)
RMSE_Test_lm = round(c(RMSE_test_lm_m,RMSE_test_lm_m_m,RMSE_test_lm_m_sd),3)

Tabla_lm = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_lm,RMSE_Test_lm))
Tabla_lm
```



### 2. KNN

```{r}
head(Train_M_Dataset )
```

#### Total Accidentes

```{r}
library(caret)
```

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_knn_fit_m = caret::train(TOTAL_ACCIDENTES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "knn", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
summary(caret_knn_fit_m)
```

```{r}
caret_knn_fit_m
```

Calculo MSE y RMSE para los datos de entrenamiento

```{r}
y_tr_pred_knn_m<-predict(caret_knn_fit_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_knn_m<-mean((Train_M_Dataset $TOTAL_ACCIDENTES-y_tr_pred_knn_m)^2) # calcula el mse de entrenamiento
RMSE_tr_knn_m = sqrt(mse_tr_knn_m)
mse_tr_knn_m
RMSE_tr_knn_m
```

Calculo MSE y RMSE para los datos de validación 

```{r}
y_test_pred_knn_m<-predict(caret_knn_fit_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_knn_m<-mean((Test_M_Dataset$TOTAL_ACCIDENTES-y_test_pred_knn_m)^2) # calcula el mse de entrenamiento
RMSE_test_knn_m = sqrt(mse_test_knn_m)
mse_test_knn_m
RMSE_test_knn_m
```

Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_knn_m,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_knn_m,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```

#### Accidentes graves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_knn_fit_m_m = caret::train(ACCIDENTES_GRAVES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "knn", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_knn_fit_m_m)
```

```{r}
caret_knn_fit_m_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_knn_m_m<-predict(caret_knn_fit_m_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_knn_m_m<-mean((Train_M_Dataset $ACCIDENTES_GRAVES-y_tr_pred_knn_m_m)^2) # calcula el mse de entrenamiento
RMSE_tr_knn_m_m = sqrt(mse_tr_knn_m_m)
mse_tr_knn_m_m
RMSE_tr_knn_m_m
```
Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_knn_m_m<-predict(caret_knn_fit_m_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_knn_m_m<-mean((Test_M_Dataset$ACCIDENTES_GRAVES-y_test_pred_knn_m_m)^2) # calcula el mse de entrenamiento
RMSE_test_knn_m_m = sqrt(mse_test_knn_m_m)
mse_test_knn_m_m
RMSE_test_knn_m_m
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_knn_m_m,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_knn_m_m,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))
```


#### Accidentes Leves


```{r}
trcntrl = trainControl(method="cv", number=10)
caret_knn_fit_m_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "knn", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_knn_fit_m_sd)
```

```{r}
caret_knn_fit_m_sd
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_knn_m_sd<-predict(caret_knn_fit_m_sd,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_knn_m_sd<-mean((Train_M_Dataset $ACCIDENTES_LEVES-y_tr_pred_knn_m_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_knn_m_sd = sqrt(mse_tr_knn_m_sd)
mse_tr_knn_m_sd
RMSE_tr_knn_m_sd
```
Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_knn_m_sd<-predict(caret_knn_fit_m_sd,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_knn_m_sd<-mean((Test_M_Dataset$ACCIDENTES_LEVES-y_test_pred_knn_m_sd)^2) # calcula el mse de entrenamiento
RMSE_test_knn_m_sd = sqrt(mse_test_knn_m_sd)
mse_test_knn_m_sd
RMSE_test_knn_m_sd
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_knn_m_sd,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_knn_m_sd,
            name='Modelo knn',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```


#### Resumen Modelos KNN para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes graves","Accidentes leves")
RMSE_Train_knn = round(c(RMSE_tr_knn_m,RMSE_tr_knn_m_m,RMSE_tr_knn_m_sd), 3)
RMSE_Test_knn = round(c(RMSE_test_knn_m,RMSE_test_knn_m_m,RMSE_test_knn_m_sd),3)

Tabla_knn = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_knn,RMSE_Test_knn))
Tabla_knn
```


### 3. MODELO LINEAL GENERALIZADO


#### Total Accidentes

```{r}
glm_fit_m<-glm(TOTAL_ACCIDENTES∼Ano_Base+MES+Feriados, data=Train_M_Dataset , family = "poisson")
summary(glm_fit_m)
```

```{r}
glm_fit_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_glm_m<-predict(glm_fit_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados")],type="response")
mse_tr_glm_m<-mean((Train_M_Dataset $TOTAL_ACCIDENTES-y_tr_pred_glm_m)^2) # calcula el mse de entrenamiento
RMSE_tr_glm_m = sqrt(mse_tr_glm_m)
mse_tr_glm_m
RMSE_tr_glm_m
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_glm_m<-predict(glm_fit_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")],type="response")
mse_test_glm_m<-mean((Train_M_Dataset $TOTAL_ACCIDENTES-y_test_pred_glm_m)^2) # calcula el mse de entrenamiento
```
```{r}
RMSE_test_glm_m = sqrt(mse_test_glm_m)
mse_test_glm_m
RMSE_test_glm_m
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_glm_m,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_glm_m,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```


#### Accidentes Graves

```{r}
glm_fit_m_m<-glm(ACCIDENTES_GRAVES∼Ano_Base+MES+Feriados, data=Train_M_Dataset , family = "poisson")
summary(glm_fit_m_m)
```

```{r}
glm_fit_m_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_glm_m_m<-predict(glm_fit_m_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados","Freq_viernes","Feriados_Lunes")],type="response")
mse_tr_glm_m_m<-mean((Train_M_Dataset $ACCIDENTES_GRAVES-y_tr_pred_glm_m_m)^2) # calcula el mse de entrenamiento
RMSE_tr_glm_m_m = sqrt(mse_tr_glm_m_m)
mse_tr_glm_m_m
RMSE_tr_glm_m_m
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_glm_m_m<-predict(glm_fit_m_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")],type="response")
mse_test_glm_m_m<-mean((Train_M_Dataset $ACCIDENTES_GRAVES-y_test_pred_glm_m_m)^2) # calcula el mse de entrenamiento
```
```{r}
RMSE_test_glm_m_m = sqrt(mse_test_glm_m_m)
mse_test_glm_m_m
RMSE_test_glm_m_m
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_glm_m_m,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_glm_m_m,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))
```



#### Accidentes Leves

```{r}
glm_fit_m_sd<-glm(ACCIDENTES_LEVES∼Ano_Base+MES+Feriados, data=Train_M_Dataset , family = "poisson")
summary(glm_fit_m_sd)
```

```{r}
glm_fit_m_sd
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_glm_m_sd<-predict(glm_fit_m_sd,Train_M_Dataset [,c("Ano_Base","MES","Feriados")],type="response")
mse_tr_glm_m_sd<-mean((Train_M_Dataset $ACCIDENTES_LEVES-y_tr_pred_glm_m_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_glm_m_sd = sqrt(mse_tr_glm_m_sd)
mse_tr_glm_m_sd
RMSE_tr_glm_m_sd
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_glm_m_sd<-predict(glm_fit_m_sd,Test_M_Dataset[,c("Ano_Base","MES","Feriados")],type="response")
mse_test_glm_m_sd<-mean((Train_M_Dataset $ACCIDENTES_LEVES-y_test_pred_glm_m_sd)^2) # calcula el mse de entrenamiento
```
```{r}
RMSE_test_glm_m_sd = sqrt(mse_test_glm_m_sd)
mse_test_glm_m_sd
RMSE_test_glm_m_sd
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_glm_m_sd,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))
```

Gráfica serie 2018  
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_glm_m_sd,
            name='Modelo glm',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total Accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))
```
#### REsumen Modelos Regresión lineal generalizado para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes Graves","Accidentes Leves")
RMSE_Train_glm = round(c(RMSE_tr_glm_m,RMSE_tr_glm_m_m,RMSE_tr_glm_m_sd), 3)
RMSE_Test_glm = round(c(RMSE_test_glm_m,RMSE_test_glm_m_m,RMSE_test_glm_m_sd),3)

Tabla_glm = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_glm,RMSE_Test_glm))
Tabla_glm
```




### 4. ARBOLES DE REGRESION

#### Total Accidentes

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_tree_fit_m = caret::train(TOTAL_ACCIDENTES∼Ano_Base+MES+Feriados,data=Train_M_Dataset ,
                              method = "rpart", trControl = trcntrl,
                      parms = list(split = "gini"),
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```



```{r}
caret_tree_fit_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_tree_m<-predict(caret_tree_fit_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_tree_m<-mean((Train_M_Dataset $TOTAL_ACCIDENTES-y_tr_pred_tree_m)^2) # calcula el mse de entrenamiento
RMSE_tr_tree_m = sqrt(mse_tr_tree_m)
mse_tr_tree_m
RMSE_tr_tree_m
```
Calculo MSE y RMSE para los datos de validación 
```{r}
  y_test_pred_tree_m<-predict(caret_tree_fit_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_tree_m<-mean((Test_M_Dataset$TOTAL_ACCIDENTES-y_test_pred_tree_m)^2) # calcula el mse de entrenamiento
RMSE_test_tree_m = sqrt(mse_test_tree_m)
mse_test_tree_m
RMSE_test_tree_m
```
Predicción en la muestra  
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_tree_m,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_tree_m,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```


#### Accidentes graves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_tree_fit_m_m = caret::train(ACCIDENTES_GRAVES∼Ano_Base+MES+Feriados,data=Train_M_Dataset ,
                              method = "rpart", trControl = trcntrl,
                      parms = list(split = "gini"),
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
caret_tree_fit_m_m
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_tree_m_m<-predict(caret_tree_fit_m_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_tree_m_m<-mean((Train_M_Dataset $ACCIDENTES_GRAVES-y_tr_pred_tree_m_m)^2) # calcula el mse de entrenamiento
RMSE_tr_tree_m_m = sqrt(mse_tr_tree_m_m)
mse_tr_tree_m_m
RMSE_tr_tree_m_m
```
Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_tree_m_m<-predict(caret_tree_fit_m_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_tree_m_m<-mean((Test_M_Dataset$ACCIDENTES_GRAVES-y_test_pred_tree_m_m)^2) # calcula el mse de entrenamiento
RMSE_test_tree_m_m = sqrt(mse_test_tree_m_m)
mse_test_tree_m_m
RMSE_test_tree_m_m
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_tree_m_m,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_tree_m_m,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```



#### Accidentes leves

```{r}
trcntrl = trainControl(method="cv", number=10)
caret_tree_fit_m_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+MES+Feriados,data=Train_M_Dataset ,
                              method = "rpart", trControl = trcntrl,
                      parms = list(split = "gini"),
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```

```{r}
caret_tree_fit_m_sd
```
Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_tree_m_sd<-predict(caret_tree_fit_m_sd,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_tree_m_sd<-mean((Train_M_Dataset $ACCIDENTES_LEVES-y_tr_pred_tree_m_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_tree_m_sd = sqrt(mse_tr_tree_m_sd)
mse_tr_tree_m_sd
RMSE_tr_tree_m_sd
```
Calculo MSE y RMSE para los datos de validación 
```{r}
  y_test_pred_tree_m_sd<-predict(caret_tree_fit_m_sd,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_tree_m_sd<-mean((Test_M_Dataset$ACCIDENTES_LEVES-y_test_pred_tree_m_sd)^2) # calcula el mse de entrenamiento
RMSE_test_tree_m_sd = sqrt(mse_test_tree_m_sd)
mse_test_tree_m_sd
RMSE_test_tree_m_sd
```
Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_tree_m_sd,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))
```
Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_tree_m_sd,
            name='Modelo tree',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))
```

#### Resumen Modelos Árboles de regresión para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Accidentes Graves","Accidentes Leves")
RMSE_Train_tree = round(c(RMSE_tr_tree_m,RMSE_tr_tree_m_m,RMSE_tr_tree_m_sd), 3)
RMSE_Test_tree = round(c(RMSE_test_tree_m,RMSE_test_tree_m_m,RMSE_test_tree_m_sd),3)

Tabla_tree = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_tree,RMSE_Test_tree))
Tabla_tree
```

### 5. BOSQUE ALEATORIO


#### Total Accidentes

```{r Random_Forest_TOTAL_ACCIDENTES}
trcntrl = trainControl(method="cv", number=10)
caret_rf_fit_s = caret::train(TOTAL_ACCIDENTES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "rf", trControl = trcntrl,
                      prox=TRUE,allowParallel=TRUE)
```


```{r}
summary(caret_rf_fit_s)
```

```{r}
caret_rf_fit_s
```

```{r}
plot(caret_rf_fit_s)
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_rf_s<-predict(caret_rf_fit_s,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_rf_s<-mean((Train_M_Dataset $TOTAL_ACCIDENTES-y_tr_pred_rf_s)^2) # calcula el mse de entrenamiento
RMSE_tr_rf_s = sqrt(mse_tr_rf_s)
mse_tr_rf_s
RMSE_tr_rf_s
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_rf_s<-predict(caret_rf_fit_s,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_rf_s<-mean((Test_M_Dataset$TOTAL_ACCIDENTES-y_test_pred_rf_s)^2) # calcula el mse de entrenamiento
RMSE_test_rf_s = sqrt(mse_test_rf_s)
mse_test_rf_s
RMSE_test_rf_s
```


Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_rf_s,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))
```


Gráfica serie 2018 

```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~TOTAL_ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y=  ~y_test_pred_rf_s,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.4))

```


#### Accidentes Graves

```{r Random_Forest_ACCIDENTES_GRAVES}
trcntrl = trainControl(method="cv", number=10)
caret_rf_fit_s_m = caret::train(ACCIDENTES_GRAVES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "rf", trControl = trcntrl,
                      prox=TRUE,allowParallel=TRUE)
```


```{r}
summary(caret_rf_fit_s_m)
```

```{r}
caret_rf_fit_s_m
```

```{r}
plot(caret_rf_fit_s_m)
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_rf_s_m<-predict(caret_rf_fit_s_m,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_rf_s_m<-mean((Train_M_Dataset $ACCIDENTES_GRAVES-y_tr_pred_rf_s_m)^2) # calcula el mse de entrenamiento
RMSE_tr_rf_s_m = sqrt(mse_tr_rf_s_m)
mse_tr_rf_s_m
RMSE_tr_rf_s_m
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_rf_s_m<-predict(caret_rf_fit_s_m,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_rf_s_m<-mean((Test_M_Dataset$ACCIDENTES_GRAVES-y_test_pred_rf_s_m)^2) # calcula el mse de entrenamiento
RMSE_test_rf_s_m = sqrt(mse_test_rf_s_m)
mse_test_rf_s_m
RMSE_test_rf_s_m
```

Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_rf_s_m,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_GRAVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_rf_s_m,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes graves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes graves"),
         legend = list(x = 1, y = 0.4))

```

#### Accidentes Leves

```{r Random_Forest_ACCIDENTES_LEVES}
trcntrl = trainControl(method="cv", number=10)
caret_rf_fit_s_sd = caret::train(ACCIDENTES_LEVES∼Ano_Base+MES+Feriados, data=Train_M_Dataset ,
                      method = "rf", trControl = trcntrl,
                      prox=TRUE,allowParallel=TRUE)
```


```{r}
summary(caret_rf_fit_s_sd)
```

```{r}
caret_rf_fit_s_sd
```

```{r}
plot(caret_rf_fit_s_sd)
```

Calculo MSE y RMSE para los datos de entrenamiento 
```{r}
y_tr_pred_rf_s_sd<-predict(caret_rf_fit_s_sd,Train_M_Dataset [,c("Ano_Base","MES","Feriados")])
mse_tr_rf_s_sd<-mean((Train_M_Dataset $ACCIDENTES_LEVES-y_tr_pred_rf_s_sd)^2) # calcula el mse de entrenamiento
RMSE_tr_rf_s_sd = sqrt(mse_tr_rf_s_sd)
mse_tr_rf_s_sd
RMSE_tr_rf_s_sd
```

Calculo MSE y RMSE para los datos de validación 
```{r}
y_test_pred_rf_s_sd<-predict(caret_rf_fit_s_sd,Test_M_Dataset[,c("Ano_Base","MES","Feriados")])
mse_test_rf_s_sd<-mean((Test_M_Dataset$ACCIDENTES_LEVES-y_test_pred_rf_s_sd)^2) # calcula el mse de entrenamiento
RMSE_test_rf_s_sd = sqrt(mse_test_rf_s_sd)
mse_test_rf_s_sd
RMSE_test_rf_s_sd
```


Predicción en la muestra 
```{r}
plot_ly (data=Train_M_Dataset ,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_tr_pred_rf_s_sd,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))
```

Gráfica serie 2018 
```{r}
plot_ly (data=Test_M_Dataset,
         x = ~Ano_mes,
         y = ~ACCIDENTES_LEVES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~y_test_pred_rf_s_sd,
            name='Modelo rf',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  layout(title='Total accidentes leves',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes leves"),
         legend = list(x = 1, y = 0.4))

```


#### Resumen Modelos Random Forest para los diferentes tipos de accidente

```{r}
Tipo_de_accidentes= c("Total Accidentes","Total Graves","Total Leves")
RMSE_Train_rf = round(c(RMSE_tr_rf_s,RMSE_tr_rf_s_m,RMSE_tr_rf_s_sd), 3)
RMSE_Test_rf = round(c(RMSE_test_rf_s,RMSE_test_rf_s_m,RMSE_test_rf_s_sd),3)

Tabla_rf = data.frame (cbind(Tipo_de_accidentes,RMSE_Train_rf,RMSE_Test_rf))
Tabla_rf
```



## ELECCION DEL MODELO


### 1. Elección del modelo para el total de accidentes

Comparación en el entrenamiento
```{r}
comparacion_tr<-data.frame(Ano_mes=Total_Dataset_Freq_M$Ano_mes[Total_Dataset_Freq_M$Ano_mes<="2017_12"],
                          ACCIDENTES=Total_Dataset_Freq_M$TOTAL_ACCIDENTES[Total_Dataset_Freq_M$Ano_mes<="2017_12"],
                           lm= y_tr_pred_lm_m, 
                           knn= y_tr_pred_knn_m, 
                           glm=y_tr_pred_glm_m ,
                           arbol=y_tr_pred_tree_m,
                           rf=y_tr_pred_rf_s)
```

```{r}
plot_ly (data=comparacion_tr,
         x = ~Ano_mes,
         y = ~ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
    add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Total Accidentes (Entrenamiento)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Unidades"),
         legend = list(x = 1, y = 0.9))
```


Comparación en la validación
```{r}
comparacion_vl<-data.frame(Ano_mes=Test_M_Dataset$Ano_mes,
                          ACCIDENTES=Test_M_Dataset$TOTAL_ACCIDENTES,
                           lm= y_test_pred_lm_m, 
                           knn= y_test_pred_knn_m, 
                           glm=y_test_pred_glm_m ,
                           arbol=y_test_pred_tree_m,
                           rf=y_test_pred_rf_s)
```

```{r}
plot_ly (data=comparacion_vl,
         x = ~Ano_mes,
         y = ~ACCIDENTES,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
  add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Total Accidentes (Validación)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```




#### Comparación con los RMSE:

```{r}
Entrenamiento<-round(c(RMSE_tr_lm_m,RMSE_tr_knn_m,RMSE_tr_glm_m,RMSE_tr_tree_m,RMSE_tr_rf_s),3) 
Validacion<-round(c(RMSE_test_lm_m,RMSE_test_knn_m,RMSE_test_glm_m,RMSE_test_tree_m,RMSE_test_rf_s),3) 
nombres<-c("lm","knn","glm","árbol","bosque")
ResultadosRMSE<-data.frame(Entrenamiento=Entrenamiento,Validacion=Validacion)
rownames(ResultadosRMSE)<-nombres
```

Cálculo de la variación

```{r}
ResultadosRMSE$Por_variacion<-((ResultadosRMSE$Validacion-ResultadosRMSE$Entrenamiento)/ResultadosRMSE$Entrenamiento)*100
ResultadosRMSE
```


### 2. Elección del modelo para Accidentes graves

Comparación en el entrenamiento
```{r}
comparacion_tr<-data.frame(Ano_mes=Total_Dataset_Freq_M$Ano_mes[Total_Dataset_Freq_M$Ano_mes<="2017_12"],
                          ACCIDENTESG=Total_Dataset_Freq_M$ACCIDENTES_GRAVES[Total_Dataset_Freq_M$Ano_mes<="2017_12"],
                           lm= y_tr_pred_lm_m_m, 
                           knn= y_tr_pred_knn_m_m, 
                           glm=y_tr_pred_glm_m_m,
                           arbol=y_tr_pred_tree_m_m,
                           rf=y_tr_pred_rf_s_m)
```

```{r}
plot_ly (data=comparacion_tr,
         x = ~Ano_mes,
         y = ~ACCIDENTESG,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbo',
            line=list(width=1,color="green"))%>%
    add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes graves (Entrenamiento)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```


Comparación en la validación
```{r}
comparacion_vl<-data.frame(Ano_mes=Test_M_Dataset$Ano_mes,
                          ACCIDENTESG=Test_M_Dataset$ACCIDENTES_GRAVES,
                           lm= y_test_pred_lm_m_m, 
                           knn= y_test_pred_knn_m_m, 
                           glm=y_test_pred_glm_m_m,
                           arbol=y_test_pred_tree_m_m,
                           rf=y_test_pred_rf_s_m)
```

```{r}
plot_ly (data=comparacion_vl,
         x = ~Ano_mes,
         y = ~ACCIDENTESG,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
  add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes graves (Validación)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```



#### Comparación con los RMSE:

```{r}
Entrenamiento<-round(c(RMSE_tr_lm_m_m,RMSE_tr_knn_m_m,RMSE_tr_glm_m_m,RMSE_tr_tree_m_m,RMSE_tr_rf_s_m),3) 
Validacion<-round(c(RMSE_test_lm_m_m,RMSE_test_knn_m_m,RMSE_test_glm_m_m,RMSE_test_tree_m_m,RMSE_test_rf_s_m),3) 
nombres<-c("lm","knn","glm","árbol","bosque")
ResultadosRMSE<-data.frame(Entrenamiento=Entrenamiento,Validacion=Validacion)
rownames(ResultadosRMSE)<-nombres
```

Cálculo de la variación

```{r}
ResultadosRMSE$Variacion<-((ResultadosRMSE$Validacion-ResultadosRMSE$Entrenamiento)/ResultadosRMSE$Entrenamiento)*100
ResultadosRMSE
```

### 3. Elección del modelo para Accidentes leves

Comparación en el entrenamiento
```{r}
comparacion_tr<-data.frame(Ano_mes=Total_Dataset_Freq_M$Ano_mes[Total_Dataset_Freq_M$Ano_mes<="2017_12"],
                          ACCIDENTESL=Total_Dataset_Freq_M$ACCIDENTES_LEVES[Total_Dataset_Freq_M$Ano_mes<="2017_12"],
                           lm= y_tr_pred_lm_m_sd, 
                           knn= y_tr_pred_knn_m_sd, 
                           glm=y_tr_pred_glm_m_sd,
                           arbol=y_tr_pred_tree_m_sd,
                           rf=y_tr_pred_rf_s_sd)
```

```{r}
plot_ly (data=comparacion_tr,
         x = ~Ano_mes,
         y = ~ACCIDENTESL,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbo',
            line=list(width=1,color="green"))%>%
    add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes leves (Entrenamiento)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```


Comparación en la validación
```{r}
comparacion_vl<-data.frame(Ano_mes=Test_M_Dataset$Ano_mes,
                          ACCIDENTESL=Test_M_Dataset$ACCIDENTES_GRAVES,
                           lm= y_test_pred_lm_m_sd, 
                           knn= y_test_pred_knn_m_sd, 
                           glm=y_test_pred_glm_m_sd,
                           arbol=y_test_pred_tree_m_sd,
                           rf=y_test_pred_rf_s_sd)
```

```{r}
plot_ly (data=comparacion_vl,
         x = ~Ano_mes,
         y = ~ACCIDENTESL,
         type = "scatter" ,mode = "lines",
         name='Real',
         line=list(width=1,color='rgb(205, 12, 24)'))%>%
  add_trace(y= ~lm,
            name='lm',
            line=list(width=1,color= "blue"))%>%
  add_trace(y= ~knn,
            name='knn',
            line=list(width=1,color="red"))%>%
  add_trace(y= ~glm,
            name='Modelo Poisson',
            line=list(width=1,color='rgb(22, 96, 167)'))%>%
  add_trace(y= ~arbol,
            name='Árbol',
            line=list(width=1,color="green"))%>%
  add_trace(y= ~rf,
            name='Bosque',
            line=list(width=1,color='rgb(255, 51, 153)'))%>%
  layout(title='Accidentes lesves (Validación)',
         xaxis=list(title="Fecha"),
         yaxis=list(title="Accidentes"),
         legend = list(x = 1, y = 0.9))
```


#### Comparación con los RMSE:

```{r}
Entrenamiento<-round(c(RMSE_tr_lm_m_sd,RMSE_tr_knn_m_sd,RMSE_tr_glm_m_sd,RMSE_tr_tree_m_sd,RMSE_tr_rf_s_sd),3) 
Validacion<-round(c(RMSE_test_lm_m_sd,RMSE_test_knn_m_sd,RMSE_test_glm_m_sd,RMSE_test_tree_m_sd,RMSE_test_rf_s_sd),3)
nombres<-c("lm","knn","glm","árbol","bosque")
ResultadosRMSE<-data.frame(Entrenamiento=Entrenamiento,Validacion=Validacion)
rownames(ResultadosRMSE)<-nombres
```

Cálculo de la variación

```{r}
ResultadosRMSE$Variacion<-((ResultadosRMSE$Validacion-ResultadosRMSE$Entrenamiento)/ResultadosRMSE$Entrenamiento)*100
ResultadosRMSE
```

#### Modelos elegidos

Teniendo como criterio el mínimo RMSE en la muestra de validación se eligen los siguientes modelos: 

##### Modelo de regresión lineal generalizado para predición de Total Accidentes 

Se ajusta el modelo con todos los datos desde el 01-01-2014 al 31-12-2018

```{r}
glm_fit_m_final<-glm(TOTAL_ACCIDENTES∼Ano_Base+MES+Feriados, data=Total_Dataset_Freq_M , family = "poisson")

```

```{r}
summary(glm_fit_m_final)
```


Se guardan el modelo en un objeto de r
```{r}
saveRDS(glm_fit_m_final,"../Modelos/Prediccion_Total_Mensual.rds")
```


```{r}
Modelo_Total_mensual<-readRDS(file="../Modelos/Prediccion_Total_Mensual.rds")
```


##### Modelo de regresión lineal generalizado para predición de Accidentes Graves

Se ajusta el modelo con todos los datos desde el 01-01-2014 al 31-12-2018
```{r}
glm_fit_m_m_final<-glm(ACCIDENTES_GRAVES∼Ano_Base+MES+Feriados, data=Total_Dataset_Freq_M, family = "poisson")

```

```{r}
summary(glm_fit_m_m_final)
```


Se guardan el modelo en un objeto de r
```{r}
saveRDS(glm_fit_m_m_final,"../Modelos/Prediccion_Grave_Mensual.rds")
```


```{r}
Modelo_Grave_mensual<-readRDS(file="../Modelos/Prediccion_Grave_Mensual.rds")
```


##### Modelo de regresión lineal para predición de Accidentes Leves

Se ajusta el modelo con todos los datos desde el 01-01-2014 al 31-12-2018
```{r}
trcntrl = trainControl(method="cv", number=10)
caret_lm_fit_m_sd_final = caret::train(ACCIDENTES_LEVES∼Ano_Base+MES+Feriados, data=Total_Dataset_Freq_M ,
                      method = "lm", trControl = trcntrl,
                      preProcess=c("center", "scale"),
                      tuneLength = 10)
```


```{r}
summary(caret_lm_fit_m_sd_final)
```



Se guardan el modelo en un objeto de r
```{r}
saveRDS(caret_lm_fit_m_sd_final,"../Modelos/Prediccion_leves_Mensual.rds")
```


```{r}
Modelo_leves_mensual<-readRDS(file="../Modelos/Prediccion_leves_Mensual.rds")
```

##### Datos para pronóstico 

Se oganizan los datos necesarios para el pronóstico de los accidentes en los años 2019, 2020 y 2021

Importación de los datos
```{r}
load("../data/Dias_Especiales_Mensuales.Rda")
```


```{r}
datos_pronostico_mensual<-Dias_Especiales_Mensuales[,c("ANO","Ano_Base","MES","Feriados")]
```

```{r}
datos_pronostico_mensual$ANO1 <- datos_pronostico_mensual$ANO
datos_pronostico_mensual$MES1 <- datos_pronostico_mensual$MES
library(dplyr)
datos_pronostico_mensual<-unite_(datos_pronostico_mensual, "Ano_Mes", c("ANO1","MES1"))
```


Predicción del Total de accidentes con el modelo de regresión lineal generalizado

```{r}
datos_pronostico_mensual$prediccion_Total_m<-predict(Modelo_Total_mensual,datos_pronostico_mensual[,c("ANO","Ano_Base","MES","Feriados")],type="response")
```

Predicción de accidentes graves con el modelo de regresión lineal generalizado

```{r}
datos_pronostico_mensual$prediccion_Graves_m<-predict(Modelo_Grave_mensual,datos_pronostico_mensual[,c("ANO","Ano_Base","MES","Feriados")],type="response")
```


Predicción de accidentes leves con el modelo de regresión lineal

```{r}
datos_pronostico_mensual$prediccion_Leves_m<-predict(Modelo_leves_mensual,datos_pronostico_mensual[,c("ANO","Ano_Base","MES","Feriados")])
```

Se guardan los datos de pronóstico en un objeto de r
```{r}
save(datos_pronostico_mensual,file="../Modelos/datos_pronostico_mensual.Rda")
```
